For #54: more architecture refining
