+++ src/main/java/com/artipie/maven/ChecksumAttribute.java
@@ -41,7 +41,7 @@ import org.slf4j.LoggerFactory;
  * Encapsulates checksum file logic.
  * Checksum files are created by Maven libarary.
  * @since 0.1
- * @deprecated Use {@link com.artipie.maven.artifact.checksum.Checksum} implementations instead
+ * @deprecated Use {@link com.artipie.maven.file.checksum.Checksum} implementations instead
  */
 @Deprecated
 public final class ChecksumAttribute {
