+++ src/main/java/com/artipie/maven/metadata/Updated.java
@@ -1,0 +1,10 @@
+package com.artipie.maven.metadata;
+
+/**
+ * Updated metadata.
+ *
+ * Reads all metadata information and updates it in the repository.
+ */
+public interface Updated extends Metadata{
+
+}
